+++ src/_local_modules/flux/actions/block-height.js
@@ -1,0 +1,7 @@
+export function updateCurrent (current) {
+  return { type: 'BLOCK_HEIGHT-UPDATE_CURRENT', payload: current }
+}
+
+export function updateTop (top) {
+  return { type: 'BLOCK_HEIGHT-UPDATE_TOP', payload: top }
+}