+++ src/_local_modules/flux/actions/index.js
@@ -1,0 +1,7 @@
+import { bindActionCreators } from 'redux'
+import * as flux from '../'
+import * as blockHeight from './block-height'
+
+export default {
+  blockHeight: bindActionCreators(blockHeight, flux.store.dispatch)
+}