+++ src/_local_modules/flux/reducers/index.js
@@ -1,0 +1,5 @@
+import blockHeight from './block-height'
+
+export default {
+  blockHeight
+}