+++ src/_local_modules/flux/selectors/block-height-percent.js
@@ -1,0 +1,9 @@
+import { createSelector } from 'reselect'
+
+export const blockHeightPercentSelector = createSelector(
+  [
+    state => state.blockHeight.current,
+    state => state.blockHeight.top
+  ],
+  (current, top) => (current / top) * 100
+)