+++ src/index.js
@@ -12,23 +12,21 @@ console.dir = function () {
   remoteConsole.dir.apply(remoteConsole, arguments)
 }
 
+console.error =function () {
+  remoteConsole.error.apply(remoteConsole, arguments)
+}
+
 // annoying popup an error box instead of silent console output
 window.onerror = function (message, url, lineno, colno, error) {
   console.error(error)
+  console.error(error.stack)
   window.alert('ERROR: \n' + message + ' (' + lineno + ')' + '\n:' + error.stack)
 }
 
 require('./stores/payments-store')
 
-var appGUI = require('./ui/index.react')
-window.onload = function () {
-  appGUI.renderApp()
-}
+document.addEventListener('DOMContentLoaded', function () {
+  require('./ui/index.react').renderApp()
+})
 
-var blockManager = require('./lib/block-manager')
-
-setTimeout(function () {
-  blockManager.start().on('error', function (err) {
-    if (err) window.alert('ERROR: \n' + err)
-  })
-}, 1000)
+require('./startup').run()