+++ src/_local_modules/flux/actions/index.js
@@ -1,7 +1,9 @@
 import { bindActionCreators } from 'redux'
 import * as flux from '../'
 import * as blockHeight from './block-height'
+import * as keys from './keys'
 
 export default {
-  blockHeight: bindActionCreators(blockHeight, flux.store.dispatch)
+  blockHeight: bindActionCreators(blockHeight, flux.store.dispatch),
+  keys: bindActionCreators(keys, flux.store.dispatch)
 }