+++ src/_local_modules/flux/actions/keys.js
@@ -1,0 +1,11 @@
+import * as keydb from '#keydb'
+
+export function load () {
+  return (dispatch, getState) => {
+    let keys = keydb.loadAllSync()
+    dispatch({
+      type: 'KEYS_LOAD',
+      payload: keys
+    })
+  }
+}