+++ src/_local_modules/flux/reducers/index.js
@@ -1,5 +1,4 @@
 import blockHeight from './block-height'
+import keys from './keys'
 
-export default {
-  blockHeight
-}
+export default { blockHeight, keys }