+++ src/_local_modules/flux/reducers/keys.js
@@ -1,0 +1,9 @@
+
+export default function keysReducer (state = [], action) {
+  switch (action.type) {
+    case 'KEYS_LOAD':
+      return action.payload
+    default:
+      return state
+  }
+}