+++ src/main/main-window.js
@@ -4,9 +4,9 @@ var window = require('electron-window')
 function initAndShow (settings, callback) {
   var mainWindow = window.createWindow({
     width: 1000,
-    height: 435,
+    height: 465,
     'min-width': 400,
-    'min-height': 415,
+    'min-height': 445,
     'web-preferences': {
       // mama don't need CORs now
       'web-security': false