+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - moved from `localStorage` to `obsidian/keys.db` to store keys
 - added a progress footer
 - started conversion from vanilla flux to redux
+- added tabs (stealth keys view)
 
 0.4.0 / 2015-07-06
 ------------------